Fix missing dist files for CDN usage
Hi, thank you for this fine package.

It seems version 0.2.5 removed the `dist` folder from NPM. This means you can no longer use services such as https://unpkg.com with this package. For instance:

- https://unpkg.com/js-video-url-parser@0.2.5/dist/jsVideoUrlParser.min.js (404)
- https://unpkg.com/js-video-url-parser@0.2.4/dist/jsVideoUrlParser.min.js (all good)

One way to fix this, is to add the `dist` folder to the `files` array in package.json to ensure it's included. There is a bit about at the bottom of this page unpkg.com/